pyproject.toml
[project]
name = "langgraph-course-v2"
version = "0.1.0"
description = ""
authors = [
    {name = "leonardo-gurdulich",email = "leonardo.gurdulich@halopowered.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily-python (>=0.7.13,<0.8.0)",
    "chromadb (>=1.3.5,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "langchain-google-genai (>=3.1.0,<4.0.0)",
    "langchain-chroma (>=1.0.0,<2.0.0)",
    "langchain-tavily (>=0.2.13,<0.3.0)",
    "langchain-classic (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
